core: Move build_select() to select.rs
