+++ core/translate/mod.rs
@@ -15,7 +15,8 @@ use crate::types::{OwnedRecord, OwnedValue};
 use crate::util::normalize_ident;
 use crate::vdbe::{builder::ProgramBuilder, BranchOffset, Insn, Program};
 use anyhow::Result;
-use expr::{build_select, maybe_apply_affinity, translate_expr};
+use expr::{maybe_apply_affinity, translate_expr};
+use select::build_select;
 use select::LeftJoinBookkeeping;
 use sqlite3_parser::ast::{self, Literal};
 use where_clause::{process_where, ProcessedWhereClause};
